Front-end validation for adding songs

+++ SongManager/src/main/java/com/songmanager/AddRemoveSongsController.java
@@ -5,12 +5,10 @@ import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
-import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.stage.Modality;
-import javafx.stage.Popup;
 import javafx.stage.Stage;
 import javafx.stage.WindowEvent;
 
@@ -18,7 +16,6 @@ import java.io.IOException;
 import java.net.URL;
 import java.sql.ResultSet;
 import java.time.Year;
-import java.util.Collection;
 import java.util.ResourceBundle;
 
 public class AddRemoveSongsController implements Initializable
@@ -31,6 +28,13 @@ public class AddRemoveSongsController implements Initializable
     private String yearString;
     private String recordLabelString;
     private String albumString;
+    private boolean validTitle;
+    private boolean validArtist;
+    private boolean validGenre;
+    private boolean validRecordLabel;
+    private boolean validAlbum;
+    private boolean validLength;
+
 
     //Overall Header
     @FXML private Label pageHeader;
@@ -89,13 +93,28 @@ public class AddRemoveSongsController implements Initializable
     //Edit Song Button
     @FXML private Button editSongBtn;
 
+    @FXML private Label titleError;
+    @FXML private Label artistError;
+    @FXML private Label albumError;
+    @FXML private Label genreError;
+    @FXML private Label lengthError;
+    @FXML private Label recordLabelError;
+
+    @FXML private Label editError;
+    @FXML private Label removeError;
+
     public AddRemoveSongsController()
     {
        this.dbManager = new DBManager("jdbc:mariadb://localhost:3306/songsproject", "root", "root");
+       this.validTitle = false;
+       this.validArtist = false;
+       this.validGenre = false;
+       this.validRecordLabel = false;
+       this.validAlbum = false;
     }
 
-   @FXML protected void AddSongBtnClick()
-   {
+    @FXML protected void AddSongBtnClick()
+    {
        if (isValidInput())
        {
            this.songTitleString = this.songTitle.getText().toString();
@@ -108,31 +127,48 @@ public class AddRemoveSongsController implements Initializable
            this.albumString = this.album.getText().toString();
 
            dbManager.addSong(songTitleString, artistString, genreString, songLengthString, yearString, recordLabelString, albumString);
-           feedbackLabel.setText("Added song");
+           feedbackLabel.setText("Added" + songTitleString);
            updateLists();
+           clearFields();
        }
-   }
+    }
 
     @FXML protected void RemoveSongBtnClick()
     {
         String songToRemove = removeSongList.getSelectionModel().getSelectedItem();
-        dbManager.deleteSong(songToRemove);
-        updateLists();
+        if (songToRemove != null)
+        {
+            dbManager.deleteSong(songToRemove);
+            this.removeError.setText("");
+            updateLists();
+        }
+        else
+        {
+            this.removeError.setText("Please select a song to remove");
+        }
     }
 
     @FXML protected void EditSongBtnClick(ActionEvent event)
     {
         String songToEdit = editSongList.getSelectionModel().getSelectedItem();
-        //dbManager.methodthatupdatessong
-        try
+        if (songToEdit != null)
         {
-            openEditSongView(event);
+            try
+            {
+                openEditSongView(event);
+                this.editError.setText("");
+            }
+            catch(IOException ex)
+            {
+                ex.printStackTrace();
+            }
+            updateLists();
         }
-        catch(IOException ex)
+        else
         {
-            ex.printStackTrace();
+            this.editError.setText("Please select a song to edit!");
         }
-        updateLists();
+
     }
 
 
@@ -163,9 +199,9 @@ public class AddRemoveSongsController implements Initializable
         });
     }
 
-   //Updates the ListViews on the page to correctly display the songs available
-   private void updateLists()
-   {
+    //Updates the ListViews on the page to correctly display the songs available
+    private void updateLists()
+    {
        ResultSet resultSet = dbManager.getSongs();
        try {
            removeSongList.getItems().clear();
@@ -180,52 +216,170 @@ public class AddRemoveSongsController implements Initializable
        {
             System.out.println(e);
        }
-       //System.out.println("FFFFdsffdsfsdfsdfsdF");
-   }
+    }
+
+    private boolean hasNulls()
+    {
+       boolean hasNulls = false;
+       //Check for empty values on columns with NOT NULL constraint
+       if (this.songTitle.getText().isEmpty())
+       {
+           this.titleError.setText("Please enter a title");
+           validTitle = false;
+           hasNulls = true;
+       }
+       else
+       {
+           this.titleError.setText("");
+       }
+       if (this.artistName.getText().isEmpty())
+       {
+           this.artistError.setText("Please enter an artist");
+           validArtist = false;
+           hasNulls = true;
+       }
+       else
+       {
+           this.artistError.setText("");
+       }
+       if (this.recordLabel.getText().isEmpty())
+       {
+           this.recordLabelError.setText("Please enter a record label");
+           validRecordLabel = false;
+           hasNulls = true;
+       }
+       else
+       {
+           this.recordLabelError.setText("");
+       }
+       if (this.album.getText().isEmpty())
+       {
+           this.albumError.setText("Please enter an album");
+           validAlbum = false;
+           hasNulls = true;
+       }
+       else
+       {
+           this.albumError.setText("");
+       }
+       if (this.genre.getText().isEmpty())
+       {
+           this.genreError.setText("Please enter an album");
+           validGenre = false;
+           hasNulls = true;
+       }
+       else
+       {
+            this.genreError.setText("");
+       }
+       return hasNulls;
+    }
+
+    private void clearFields()
+    {
+        this.songTitle.setText("");
+        this.artistName.setText("");
+        this.album.setText("");
+        this.genre.setText("");
+        this.hours.setText("");
+        this.minutes.setText("");
+        this.seconds.setText("");
+        this.recordLabel.setText("");
+        this.yearReleased.setValue(null);
+    }
 
     private boolean isValidInput()
     {
         boolean valid = true;
         //Check for empty values on columns with NOT NULL constraint
-        if (this.songTitle != null && this.artistName != null && this.genre != null && yearReleased != null)
+        if (!hasNulls())
         {
             //Check for proper data types
             //songTitle
             if (this.songTitle.getText().length() > 25)
             {
                 valid = false;
+                validTitle = false;
+                this.titleError.setText("Title must be 25 or fewer characters");
+            }
+            else
+            {
+                this.titleError.setText("");
             }
             //artistID
             if (this.artistName.getText().length() > 25)
             {
                 valid = false;
+                validArtist = false;
+                this.artistError.setText("Artist name must be 25 or fewer characters");
+            }
+            else
+            {
+                this.artistError.setText("");
             }
             //genreName
             if (this.genre.getText().length() > 15)
             {
                 valid = false;
+                validGenre = false;
+                this.genreError.setText("Genre name must be 25 or fewer characters");
+            }
+            else
+            {
+                this.genreError.setText("");
             }
             //songLength
-           /* try
+            try
             {
-                Duration songLength = new Duration()
-                if ()
+                int hoursInt, minutesInt, secondsInt;
+                hoursInt = Integer.parseInt(this.hours.getText());
+                minutesInt = Integer.parseInt(this.minutes.getText());
+                secondsInt = Integer.parseInt(this.seconds.getText());
+                if (hoursInt < 0 || minutesInt < 0 || minutesInt > 59 || secondsInt < 0 || secondsInt > 59)
+                {
+                    valid = false;
+                    validLength = false;
+                    this.lengthError.setText("Invalid length");
+                }
+                else
+                {
+                    this.lengthError.setText("");
+                }
             }
             //not a valid integer
             catch (Exception ex)
             {
-                System.out.println("Please enter a valid length!");
-            }*/
+                valid = false;
+                validLength = false;
+                this.lengthError.setText("Length values must be integers");
+            }
+
             //recordLabel
             if (this.recordLabel.getText().length() > 25)
             {
                 valid = false;
+                validRecordLabel = false;
+                this.recordLabelError.setText("Record label name must be 25 or fewer characters");
+            }
+            else
+            {
+                this.recordLabelError.setText("");
             }
             //album
             if (this.album.getText().length() > 25)
             {
                 valid = false;
+                this.album.setText("Album name must be 25 or fewer characters");
             }
+            else
+            {
+                this.albumError.setText("");
+            }
+        }
+        //Found nulls
+        else
+        {
+            valid = false;
         }
         return valid;
     }
